+++ service/conversations/v1/service/conversation/api_op_client.go
@@ -1,7 +1,11 @@
 // Package conversation contains auto-generated files. DO NOT MODIFY
 package conversation
 
-import "github.com/RJPearson94/twilio-sdk-go/client"
+import (
+	"github.com/RJPearson94/twilio-sdk-go/client"
+	"github.com/RJPearson94/twilio-sdk-go/service/conversations/v1/service/conversation/webhook"
+	"github.com/RJPearson94/twilio-sdk-go/service/conversations/v1/service/conversation/webhooks"
+)
 
 // Client for managing a specific conversation resource
 // See https://www.twilio.com/docs/conversations/api/conversation-resource for more details
@@ -10,6 +14,9 @@ type Client struct {
 
 	serviceSid string
 	sid        string
+
+	Webhook  func(string) *webhook.Client
+	Webhooks *webhooks.Client
 }
 
 // ClientProperties are the properties required to manage the conversation resources
@@ -25,5 +32,17 @@ func New(client *client.Client, properties ClientProperties) *Client {
 
 		serviceSid: properties.ServiceSid,
 		sid:        properties.Sid,
+
+		Webhook: func(webhookSid string) *webhook.Client {
+			return webhook.New(client, webhook.ClientProperties{
+				ConversationSid: properties.Sid,
+				ServiceSid:      properties.ServiceSid,
+				Sid:             webhookSid,
+			})
+		},
+		Webhooks: webhooks.New(client, webhooks.ClientProperties{
+			ConversationSid: properties.Sid,
+			ServiceSid:      properties.ServiceSid,
+		}),
 	}
 }